Move dns code from network package to dns package
