+++ pkg/crc/network/types.go
@@ -7,19 +7,6 @@ import (
 	"github.com/spf13/cast"
 )
 
-type NameServer struct {
-	IPAddress string
-}
-
-type SearchDomain struct {
-	Domain string
-}
-
-type ResolvFileValues struct {
-	SearchDomains []SearchDomain
-	NameServers   []NameServer
-}
-
 type Mode string
 
 const (
