+++ pkg/crc/services/dns/nameservers.go
@@ -1,4 +1,4 @@
-package network
+package dns
 
 import (
 	"fmt"
