+++ pkg/crc/services/dns/resolvtemplate.go
@@ -1,4 +1,4 @@
-package network
+package dns
 
 import (
 	"bytes"
