+++ pkg/crc/services/dns/types.go
@@ -1,0 +1,14 @@
+package dns
+
+type NameServer struct {
+	IPAddress string
+}
+
+type SearchDomain struct {
+	Domain string
+}
+
+type ResolvFileValues struct {
+	SearchDomains []SearchDomain
+	NameServers   []NameServer
+}
